refactor(js): separar arquivo ``script.js`` em módulos ``calendar.js``, ``utils.js`` e ``main.js``
