+++ index.html
@@ -72,160 +72,8 @@
       <span class="top-bar__days">Dom</span>
     </section>
 
-    <section class="calendar__week">
-      <div class="calendar__day inactive">
-        <span class="calendar__date">30</span>
-        <span class="calendar__task">2</span>
-      </div>
-      <div class="calendar__day inactive">
-        <span class="calendar__date">31</span>
-        <span class="calendar__task">4</span>
-      </div>
-      <div class="calendar__day">
-        <span class="calendar__date">1</span>
-        <span class="calendar__task">2</span>
-      </div>
-      <div class="calendar__day">
-        <span class="calendar__date">2</span>
-        <span class="calendar__task">3</span>
-      </div>
-      <div class="calendar__day">
-        <span class="calendar__date">3</span>
-        <span class="calendar__task">1</span>
-      </div>
-      <div class="calendar__day">
-        <span class="calendar__date">4</span>
-        <span class="calendar__task">2</span>
-      </div>
-      <div class="calendar__day">
-        <span class="calendar__date">5</span>
-        <span class="calendar__task">4</span>
-      </div>
-    </section>
 
-    <section class="calendar__week">
-      <div class="calendar__day">
-        <span class="calendar__date">6</span>
-        <span class="calendar__task">2</span>
-      </div>
-      <div class="calendar__day">
-        <span class="calendar__date">7</span>
-        <span class="calendar__task">3</span>
-      </div>
-      <div class="calendar__day">
-        <span class="calendar__date">8</span>
-        <span class="calendar__task">3</span>
-      </div>
-      <div class="calendar__day">
-        <span class="calendar__date">9</span>
-        <span class="calendar__task">1</span>
-      </div>
-      <div class="calendar__day">
-        <span class="calendar__date">10</span>
-        <span class="calendar__task">2</span>
-      </div>
-      <div class="calendar__day">
-        <span class="calendar__date">11</span>
-        <span class="calendar__task">2</span>
-      </div>
-      <div class="calendar__day">
-        <span class="calendar__date">12</span>
-        <span class="calendar__task">1</span>
-      </div>
-    </section>
 
-    <section class="calendar__week">
-      <div class="calendar__day">
-        <span class="calendar__date">13</span>
-        <span class="calendar__task">2</span>
-      </div>
-      <div class="calendar__day">
-        <span class="calendar__date">14</span>
-        <span class="calendar__task">2</span>
-      </div>
-      <div class="calendar__day today">
-        <span class="calendar__date">15</span>
-        <span class="calendar__task calendar__task--today"></span>
-      </div>
-      <div class="calendar__day">
-        <span class="calendar__date">16</span>
-        <span class="calendar__task">2</span>
-      </div>
-      <div class="calendar__day">
-        <span class="calendar__date">17</span>
-        <span class="calendar__task">1</span>
-      </div>
-      <div class="calendar__day">
-        <span class="calendar__date">18</span>
-        <span class="calendar__task">4</span>
-      </div>
-      <div class="calendar__day">
-        <span class="calendar__date">19</span>
-        <span class="calendar__task">2</span>
-      </div>
-    </section>
-
-    <section class="calendar__week">
-      <div class="calendar__day">
-        <span class="calendar__date">20</span>
-        <span class="calendar__task">2</span>
-      </div>
-      <div class="calendar__day">
-        <span class="calendar__date">21</span>
-        <span class="calendar__task">2</span>
-      </div>
-      <div class="calendar__day">
-        <span class="calendar__date">22</span>
-        <span class="calendar__task">1</span>
-      </div>
-      <div class="calendar__day">
-        <span class="calendar__date">23</span>
-        <span class="calendar__task">2</span>
-      </div>
-      <div class="calendar__day">
-        <span class="calendar__date">24</span>
-        <span class="calendar__task">5</span>
-      </div>
-      <div class="calendar__day">
-        <span class="calendar__date">25</span>
-        <span class="calendar__task">3</span>
-      </div>
-      <div class="calendar__day">
-        <span class="calendar__date">26</span>
-        <span class="calendar__task">1</span>
-      </div>
-    </section>
-
-    <section class="calendar__week">
-      <div class="calendar__day">
-        <span class="calendar__date">27</span>
-        <span class="calendar__task">2</span>
-      </div>
-      <div class="calendar__day">
-        <span class="calendar__date">28</span>
-        <span class="calendar__task">1</span>
-      </div>
-      <div class="calendar__day inactive">
-        <span class="calendar__date">1</span>
-        <span class="calendar__task">2</span>
-      </div>
-      <div class="calendar__day inactive">
-        <span class="calendar__date">2</span>
-        <span class="calendar__task">1</span>
-      </div>
-      <div class="calendar__day inactive">
-        <span class="calendar__date">3</span>
-        <span class="calendar__task">5</span>
-      </div>
-      <div class="calendar__day inactive">
-        <span class="calendar__date">4</span>
-        <span class="calendar__task">3</span>
-      </div>
-      <div class="calendar__day inactive">
-        <span class="calendar__date">5</span>
-        <span class="calendar__task">1</span>
-      </div>
-    </section>
   </section>
 
 </main>
@@ -260,6 +108,7 @@
     });
   });
 </script>
+<script type="module" src="./scripts/main.js"></script>
 
 </body>
 </html>
