+++ scripts/main.js
@@ -1,0 +1,4 @@
+import {renderCalendar} from './calendar.js'
+
+
+renderCalendar();
